AgentChat-Mcp-magentic-orchestration-improvements/PythonAPI/requirements.txt
# Core dependencies
flask[async]==2.3.3
flask-cors==4.0.0
pydantic>=2.8.0
uvicorn==0.31.1
fastapi>=0.110.0
starlette>=0.37.0

# MCP SDK
# MCP SDK (require version that includes streamable_http client)
mcp[cli]==1.12.4
fastmcp>=0.1.0

# Semantic Kernel for multi-agent system
# Semantic Kernel for multi-agent system (pin to version tested locally)
semantic-kernel==1.35.0

# Development tools
python-dotenv==1.0.1

# Azure OpenAI
openai>=1.30.0

# Azure Cosmos DB
azure-cosmos>=4.5.0

# Azure Blob Storage
azure-storage-blob>=12.20.0

# Azure AI Search
azure-search-documents>=11.5.0

# Azure Form Recognizer / Document Intelligence
azure-ai-formrecognizer>=3.3.0

# Text extraction libraries (optional)
chardet>=5.0.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
python-docx>=1.1.0
python-pptx>=0.6.0
openpyxl>=3.1.0

# Azure deployment (future)
gunicorn==22.0.0
azure-identity>=1.16.0
azure-keyvault-secrets>=4.8.0

# Azure Application Insights
opencensus-ext-azure>=1.1.13
opencensus-ext-flask>=0.8.0
opencensus-ext-requests>=0.8.0
opencensus-ext-logging>=0.1.1

# Azure Data Explorer
azure-kusto-data>=4.5.0
azure-kusto-ingest>=4.5.0

# Optional utilities
httpx>=0.27.0
aiohttp>=3.8.0
pydantic-settings>=2.3.0
